water_strip_app/requirements.txt
streamlit
numpy
opencv-python-headless
scikit-learn==1.6.1
Pillow